Create Core Transforms Map ParDo task
