+++ learning/katas/go/Core Transforms/Map/ParDo/go.mod
@@ -16,3 +16,12 @@
 module pardo
 
 go 1.13
+
+require (
+	github.com/apache/beam v2.20.0+incompatible
+	github.com/golang/protobuf v1.4.0 // indirect
+	github.com/googleapis/gax-go v2.0.2+incompatible // indirect
+	go.opencensus.io v0.22.3 // indirect
+	google.golang.org/api v0.21.0 // indirect
+	google.golang.org/grpc v1.28.1 // indirect
+)
